b/requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
Django==5.1.4
django-filter==24.3
django-graphql-auth==0.3.16
django-graphql-jwt==0.4.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
Rx==1.6.3
singledispatch==4.1.0
six==1.17.0
sqlparse==0.5.2
text-unidecode==1.3
typing_extensions==4.12.2
